+++ include/bp/graph.hpp
@@ -36,13 +36,13 @@ class Graph {
 
   void addNode(std::shared_ptr<GraphNode> node);
 
-  bool hasNode(size_t index) const;
+  bool hasNode(int index) const;
 
-  bool hasFactor(size_t index, BPFactorType t) const;
+  bool hasFactor(int index, BPFactorType t) const;
 
-  std::shared_ptr<GraphNode> getNode(size_t index) const;
+  std::shared_ptr<GraphNode> getNode(int index) const;
 
-  std::shared_ptr<GraphFactor> getFactor(size_t index, BPFactorType t) const;
+  std::shared_ptr<GraphFactor> getFactor(int index, BPFactorType t) const;
 
   double entropySum() const;
 
@@ -52,13 +52,13 @@ class Graph {
                          std::shared_ptr<GraphNode> node,
                          IODirection dir);
 
-  size_t iterations() const;
+  int iterations() const;
 
   void norm();
 
   void initMessages();
 
-  void spreadPriors(const std::vector<size_t> &prior_rvs);
+  void spreadPriors(const std::vector<int> &prior_rvs);
 
   void scheduledUpdate();
 
@@ -66,11 +66,11 @@ class Graph {
   std::vector<std::vector<std::shared_ptr<GraphFactor>>> schedule_factor;
 
  private:
-  size_t iter_;
+  int iter_;
   std::vector<std::shared_ptr<GraphFactor>> factors_;
   std::map<std::string, std::shared_ptr<GraphFactor>> factor_map_;
   std::vector<std::shared_ptr<GraphNode>> nodes_;
-  std::map<size_t, std::shared_ptr<GraphNode>> node_map_;
+  std::map<int, std::shared_ptr<GraphNode>> node_map_;
 };
 
 }  // end namespace bp
