+++ src/bp/graph.cpp
@@ -74,7 +74,7 @@ void Graph::printGraph() const {
 void Graph::writeNodes() const {
   std::ofstream outfile;
   outfile.open("/tmp/bp_dist.txt", std::ios::out | std::ios::app);
-  for (size_t i = 0; i < nodes_.size(); i++) {
+  for (int i = 0; i < nodes_.size(); i++) {
     outfile << nodes_.at(i)->distanceFromUndetermined();
     if (i != nodes_.size() - 1) outfile << ",";
   }
@@ -82,7 +82,7 @@ void Graph::writeNodes() const {
   outfile.close();
 
   outfile.open("/tmp/bp_bits.txt", std::ios::out | std::ios::app);
-  for (size_t i = 0; i < nodes_.size(); i++) {
+  for (int i = 0; i < nodes_.size(); i++) {
     outfile << (int)nodes_.at(i)->bit();
     if (i != nodes_.size() - 1) outfile << ",";
   }
@@ -100,20 +100,20 @@ void Graph::addNode(std::shared_ptr<GraphNode> node) {
   node_map_[node->index()] = node;
 }
 
-bool Graph::hasNode(size_t index) const {
+bool Graph::hasNode(int index) const {
   return node_map_.count(index) > 0;
 }
 
-bool Graph::hasFactor(size_t index, BPFactorType t) const {
+bool Graph::hasFactor(int index, BPFactorType t) const {
   const std::string s = GraphFactor::makeString(index, t);
   return factor_map_.count(s) > 0;
 }
 
-std::shared_ptr<GraphNode> Graph::getNode(size_t index) const {
+std::shared_ptr<GraphNode> Graph::getNode(int index) const {
   return node_map_.at(index);
 }
 
-std::shared_ptr<GraphFactor> Graph::getFactor(size_t index, BPFactorType t) const {
+std::shared_ptr<GraphFactor> Graph::getFactor(int index, BPFactorType t) const {
   const std::string s = GraphFactor::makeString(index, t);
   return factor_map_.at(s);
 }
@@ -138,7 +138,7 @@ void Graph::connectFactorNode(std::shared_ptr<GraphFactor> factor,
   node->addEdge(e);
 }
 
-size_t Graph::iterations() const { return iter_; }
+int Graph::iterations() const { return iter_; }
 
 void Graph::norm() {
   for (auto &node : nodes_) node->norm();
@@ -149,15 +149,15 @@ void Graph::initMessages() {
   for (auto &node : nodes_) node->initMessages();
 }
 
-void Graph::spreadPriors(const std::vector<size_t> &prior_rvs) {
-  for (size_t rv : prior_rvs) {
+void Graph::spreadPriors(const std::vector<int> &prior_rvs) {
+  for (int rv : prior_rvs) {
     if (hasNode(rv)) getNode(rv)->node2factor();
   }
 }
 
 void Graph::scheduledUpdate() {
   iter_++;
-  const size_t n_layers = schedule_factor.size();
+  const int n_layers = schedule_factor.size();
 
 #ifdef PRINT_DEBUG
   spdlog::info("A: starting scheduledUpdate()");
@@ -174,7 +174,7 @@ void Graph::scheduledUpdate() {
 #endif
 
   // ################# FORWARD  #################
-  for (size_t r = 0; r < n_layers; r++) {
+  for (int r = 0; r < n_layers; r++) {
     for (auto &factor : schedule_factor.at(r)) {
       factor->factor2node();
     }
@@ -199,10 +199,10 @@ void Graph::scheduledUpdate() {
 
   // ################# BACKWARD  #################
   for (int r = (int)n_layers - 1; r >= 0; r--) {
-    for (size_t i = schedule_factor.at(r).size(); i-- > 0;) {
+    for (int i = schedule_factor.at(r).size(); i-- > 0;) {
       schedule_factor.at(r).at(i)->factor2node();
     }
-    for (size_t i = schedule_variable.at(r).size(); i-- > 0;) {
+    for (int i = schedule_variable.at(r).size(); i-- > 0;) {
       schedule_variable.at(r).at(i)->node2factor(IODirection::Output);
     }
   }
